# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: ""
  version: 0.0.1
...
paths:
...
  /credentials:
    get:
      tags:
        - CredentialsService
      operationId: ListCredentials
      parameters:
        - name: purpose
          in: query
          description: Return only credentials for the specified purpose.
          schema:
            type: integer
            format: enum
        - name: maxResults
          in: query
          description: The maximum number of results per page that should be returned.
          schema:
            type: integer
            format: int32
        - name: pageToken
          in: query
          description: Opaque pagination token to go to next page based on previous query.
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListCredentialsResponse"
        default:
          description: Default error response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Status"
    post:
      tags:
        - CredentialsService
      operationId: CreateCredential
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCredentialRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Credential"
        default:
          description: Default error response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Status"
...
  /external-locations:
    get:
      tags:
        - ExternalLocationsService
      description: List external locations
      operationId: ListExternalLocations
      parameters:
        - name: maxResults
          in: query
          description: The maximum number of results per page that should be returned.
          schema:
            type: integer
            format: int32
        - name: pageToken
          in: query
          description: Opaque pagination token to go to next page based on previous query.
          schema:
            type: string
        - name: includeBrowse
          in: query
          description:
            Whether to include schemas in the response for which the principal can only access selective metadata
            for
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListExternalLocationsResponse"
        default:
          description: Default error response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Status"
    post:
      tags:
        - ExternalLocationsService
      description: Create a new external location
      operationId: CreateExternalLocation
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExternalLocationRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExternalLocation"
        default:
          description: Default error response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Status"
...
  /shares:
    get:
      tags:
        - SharesService
      description: List shares.
      operationId: ListShares
      parameters:
        - name: maxResults
          in: query
          description: The maximum number of results per page that should be returned.
          schema:
            type: integer
            format: int32
        - name: pageToken
          in: query
          description: Opaque pagination token to go to next page based on previous query.
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListSharesResponse"
        default:
          description: Default error response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Status"
    post:
      tags:
        - SharesService
      description: Create a new share.
      operationId: CreateShare
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShareRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Share"
        default:
          description: Default error response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Status"
  /shares/{name}:
    get:
      tags:
        - SharesService
      description: Get a share by name.
      operationId: GetShare
      parameters:
        - name: name
          in: path
          description: Name of the share.
          required: true
          schema:
            type: string
        - name: includeSharedData
          in: query
          description: Query for data to include in the share.
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Share"
        default:
          description: Default error response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Status"
    delete:
      tags:
        - SharesService
      description: Deletes a share.
      operationId: DeleteShare
      parameters:
        - name: name
          in: path
          description: Name of the share.
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content: {}
        default:
          description: Default error response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Status"
    patch:
      tags:
        - SharesService
      description: Update a share.
      operationId: UpdateShare
      parameters:
        - name: name
          in: path
          description: Name of the share.
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateShareRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Share"
        default:
          description: Default error response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Status"
...
components:
  schemas:
...
    CreateCredentialRequest:
      additionalProperties: false
      properties:
        application_id:
          description: The application ID of the application registration within the referenced AAD tenant.
          type: string
        msi_resource_id:
          description: Msi resource id for use with managed identity authentication
          type: string
        object_id:
          description: Object id for use with managed identity authentication
          type: string
      title: Azure Managed Identity
      type: object
    CreateExternalLocationRequest:
      additionalProperties: false
      description: Create a new external location
      properties:
        comment:
          description: User-provided free-form text description.
          type: string
        credential_name:
          description: Name of the storage credential used with this location.
          minLength: 3
          pattern: ^[a-z][0-9a-z_]*[0-9a-z]$
          type: string
        name:
          description: Name of external location.
          minLength: 3
          pattern: ^[a-z][0-9a-z_]*[0-9a-z]$
          type: string
        read_only:
          description: Indicates whether the external location is read-only.
          type: boolean
        skip_validation:
          description: Skips validation of the storage credential associated with the external location.
          type: boolean
        url:
          description: Path URL of the external location.
          pattern: ^(?:(?:[a-zA-Z][a-zA-Z\d+\-.]*):)?(?://(?:[A-Za-z0-9\-\.]+(?::\d+)?))?(/[^\?#]*)?(?:\?([^\#]*))?(?:\#(.*))?$
          type: string
      required:
        - name
        - url
        - credential_name
      title: Create External Location Request
      type: object
...
    CreateShareRequest:
      additionalProperties: false
      description:
        "Data objects can be added after creation with update.\n The caller must be a metastore admin or have the\
        \ CREATE_SHARE privilege on the metastore."
      properties:
        comment:
          description: User-provided free-form text description.
          type: string
        name:
          description: Name of the share.
          minLength: 3
          pattern: ^[a-z][0-9a-z_]*[0-9a-z]$
          type: string
      required:
        - name
      title: Creates a new share for data objects.
      type: object
...
    Credential:
      additionalProperties: false
      properties:
        application_id:
          description: The application ID of the application registration within the referenced AAD tenant.
          type: string
        msi_resource_id:
          description: Msi resource id for use with managed identity authentication
          type: string
        object_id:
          description: Object id for use with managed identity authentication
          type: string
      title: Azure Managed Identity
      type: object
...
    ExternalLocation:
      additionalProperties: false
      properties:
        browse_only:
          description:
            "Indicates whether the principal is limited to retrieving metadata\n for the associated object through\
            \ the BROWSE privilege when include_browse is enabled in the request."
          type: boolean
        comment:
          description: User-provided free-form text description.
          type: string
        created_at:
          description: Time at which this catalog was created, in epoch milliseconds.
          type: integer
        created_by:
          description: Username of catalog creator.
          type: string
        credential_id:
          description: Unique ID of the location's storage credential.
          type: string
        credential_name:
          description: Name of the storage credential used with this location.
          type: string
        external_location_id:
          type: string
        name:
          description: Name of the external location.
          type: string
        owner:
          description: The owner of the external location.
          type: string
        read_only:
          description: Indicates whether the external location is read-only.
          type: boolean
        updated_at:
          description: Time at which this catalog was last updated, in epoch milliseconds.
          type: integer
        updated_by:
          description: Username of user who last modified catalog.
          type: string
        url:
          description: Path URL of the external location.
          type: string
      required:
        - name
        - url
        - credential_name
        - read_only
        - credential_id
      title: External Location
      type: object
...
    GoogleProtobufAny:
      type: object
      properties:
        "@type":
          type: string
          description: The type of the serialized message.
      additionalProperties: true
      description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
...
    ListCredentialsResponse:
      additionalProperties: false
      properties:
        application_id:
          description: The application ID of the application registration within the referenced AAD tenant.
          type: string
        msi_resource_id:
          description: Msi resource id for use with managed identity authentication
          type: string
        object_id:
          description: Object id for use with managed identity authentication
          type: string
      title: Azure Managed Identity
      type: object
    ListExternalLocationsResponse:
      additionalProperties: false
      properties:
        browse_only:
          description:
            "Indicates whether the principal is limited to retrieving metadata\n for the associated object through\
            \ the BROWSE privilege when include_browse is enabled in the request."
          type: boolean
        comment:
          description: User-provided free-form text description.
          type: string
        created_at:
          description: Time at which this catalog was created, in epoch milliseconds.
          type: integer
        created_by:
          description: Username of catalog creator.
          type: string
        credential_id:
          description: Unique ID of the location's storage credential.
          type: string
        credential_name:
          description: Name of the storage credential used with this location.
          type: string
        external_location_id:
          type: string
        name:
          description: Name of the external location.
          type: string
        owner:
          description: The owner of the external location.
          type: string
        read_only:
          description: Indicates whether the external location is read-only.
          type: boolean
        updated_at:
          description: Time at which this catalog was last updated, in epoch milliseconds.
          type: integer
        updated_by:
          description: Username of user who last modified catalog.
          type: string
        url:
          description: Path URL of the external location.
          type: string
      required:
        - name
        - url
        - credential_name
        - read_only
        - credential_id
      title: External Location
      type: object
...
    ListSharesResponse:
      additionalProperties: false
      properties:
        added_at:
          description: The time when this data object is added to the share, in epoch milliseconds.
          type: integer
        added_by:
          description: Username of the sharer.
          type: string
        comment:
          description: A user-provided comment when adding the data object to the share.
          type: string
        data_object_type:
          description: Type of the data object.
          enum:
            - DATA_OBJECT_TYPE_UNSPECIFIED
            - TABLE
            - SCHEMA
          title: Data Object Type
          type: string
        enable_cdf:
          description: Whether to enable cdf or indicate if cdf is enabled on the shared object.
          type: boolean
        history_data_sharing_status:
          description: Whether to enable or disable sharing of data history. If not specified, the default is DISABLED.
          enum:
            - DISABLED
            - ENABLED
          title: History Status
          type: string
        name:
          description: For example, a table's fully qualified name is in the format of <catalog>.<schema>.<table>,
          title: A fully qualified name that uniquely identifies a data object.
          type: string
        partitions:
          description: Array of partitions for the shared data.
          items:
            type: string
          type: array
        shared_as:
          description:
            "If this new name is not provided, the object's original name will be used as the shared_as name.\n\
            \ The shared_as name must be unique within a share.\n For tables, the new name must follow the format of <schema>.<table>."
          title: A user-provided new name for the data object within the share.
          type: string
        start_version:
          description:
            "This allows data providers to control the lowest object version that is accessible by clients.\n If\
            \ specified, clients can query snapshots or changes for versions >= start_version.\n If not specified, clients\
            \ can only query starting from the version of the object at the time it was added to the share.\n\n NOTE: The\
            \ start_version should be <= the current version of the object."
          title: The start version associated with the object.
          type: integer
      required:
        - name
        - data_object_type
      title: Data Object
      type: object
...
    Share:
      additionalProperties: false
      properties:
        added_at:
          description: The time when this data object is added to the share, in epoch milliseconds.
          type: integer
        added_by:
          description: Username of the sharer.
          type: string
        comment:
          description: A user-provided comment when adding the data object to the share.
          type: string
        data_object_type:
          description: Type of the data object.
          enum:
            - DATA_OBJECT_TYPE_UNSPECIFIED
            - TABLE
            - SCHEMA
          title: Data Object Type
          type: string
        enable_cdf:
          description: Whether to enable cdf or indicate if cdf is enabled on the shared object.
          type: boolean
        history_data_sharing_status:
          description: Whether to enable or disable sharing of data history. If not specified, the default is DISABLED.
          enum:
            - DISABLED
            - ENABLED
          title: History Status
          type: string
        name:
          description: For example, a table's fully qualified name is in the format of <catalog>.<schema>.<table>,
          title: A fully qualified name that uniquely identifies a data object.
          type: string
        partitions:
          description: Array of partitions for the shared data.
          items:
            type: string
          type: array
        shared_as:
          description:
            "If this new name is not provided, the object's original name will be used as the shared_as name.\n\
            \ The shared_as name must be unique within a share.\n For tables, the new name must follow the format of <schema>.<table>."
          title: A user-provided new name for the data object within the share.
          type: string
        start_version:
          description:
            "This allows data providers to control the lowest object version that is accessible by clients.\n If\
            \ specified, clients can query snapshots or changes for versions >= start_version.\n If not specified, clients\
            \ can only query starting from the version of the object at the time it was added to the share.\n\n NOTE: The\
            \ start_version should be <= the current version of the object."
          title: The start version associated with the object.
          type: integer
      required:
        - name
        - data_object_type
      title: Data Object
      type: object
    Status:
      type: object
      properties:
        code:
          type: integer
          description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
          format: int32
        message:
          type: string
          description:
            A developer-facing error message, which should be in English. Any user-facing error message should
            be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the
            client.
        details:
          type: array
          items:
            $ref: "#/components/schemas/GoogleProtobufAny"
          description:
            A list of messages that carry the error details.  There is a common set of message types for APIs to
            use.
      description:
        "The `Status` type defines a logical error model that is suitable for different programming environments,
        including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three
        pieces of data: error code, error message, and error details. You can find out more about this error model and how
        to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors)."
...
    UpdateShareRequest:
      additionalProperties: false
      properties:
        added_at:
          description: The time when this data object is added to the share, in epoch milliseconds.
          type: integer
        added_by:
          description: Username of the sharer.
          type: string
        comment:
          description: A user-provided comment when adding the data object to the share.
          type: string
        data_object_type:
          description: Type of the data object.
          enum:
            - DATA_OBJECT_TYPE_UNSPECIFIED
            - TABLE
            - SCHEMA
          title: Data Object Type
          type: string
        enable_cdf:
          description: Whether to enable cdf or indicate if cdf is enabled on the shared object.
          type: boolean
        history_data_sharing_status:
          description: Whether to enable or disable sharing of data history. If not specified, the default is DISABLED.
          enum:
            - DISABLED
            - ENABLED
          title: History Status
          type: string
        name:
          description: For example, a table's fully qualified name is in the format of <catalog>.<schema>.<table>,
          title: A fully qualified name that uniquely identifies a data object.
          type: string
        partitions:
          description: Array of partitions for the shared data.
          items:
            type: string
          type: array
        shared_as:
          description:
            "If this new name is not provided, the object's original name will be used as the shared_as name.\n\
            \ The shared_as name must be unique within a share.\n For tables, the new name must follow the format of <schema>.<table>."
          title: A user-provided new name for the data object within the share.
          type: string
        start_version:
          description:
            "This allows data providers to control the lowest object version that is accessible by clients.\n If\
            \ specified, clients can query snapshots or changes for versions >= start_version.\n If not specified, clients\
            \ can only query starting from the version of the object at the time it was added to the share.\n\n NOTE: The\
            \ start_version should be <= the current version of the object."
          title: The start version associated with the object.
          type: integer
      required:
        - name
        - data_object_type
      title: Data Object
      type: object
...